Split the lookup index by type and settle its prefixes
Two files in the module carried more than the type they were named
after: RecordCacheStore held the SwiftData models alongside the store,
and CachedDatabase held the registry alongside the database wrapper.
Both split, so the module now runs one type per file.

The names had drifted into four families — Cache, Cached, Record and
Database — with two types on their own in the first and last. They
settle into the two the module actually has: Cached for what is cached,
RecordCache for the machinery around it. CacheRow becomes
RecordCacheRow, and DatabaseCacheRegistry becomes RecordCacheRegistry,
which is also the truer name: it registers a record cache, not a
database.

SeriesKey and SeriesGroups fold into CachedMetricSeries, their only
caller, taking the last unprefixed file with them. That leaves eight
files where sibling targets sit flat with as many as twelve, so the
module stays without subfolders. The file-private alias over
CachedLookupTypes.all goes too — the set it aliased is already a static
let, so the indirection bought nothing.
